+++ NEWS.md
@@ -3,6 +3,8 @@
 * ROC/AUC function added
 * Function to generate bootstrapped confidence intervals for AUC added.
 * Various S3 methods for joint objects added.
+* Relaxation of (quite stringent) model calls.
+* Extraction of conditional distribution of random effects.
 
 # gmvjoint version 0.1.0
 * First release